+++ changelog.md
@@ -267,3 +267,15 @@ This is optional and allows consumers to provide further topic names to subsribe
 ## 3.44 2026-05-12
 
 `ecs/ec2_capacity_provider` and `ecs/ec2_capacity_provider_abs` output ASG arn.
+
+## 3.45 2026-05-29
+
+`ecs/ec2_capacity_provider` and `ecs/ec2_capacity_provider_abs` default AMI updated to AmazonLinux 2023.
+
+Also changes from using `data "aws_ami"` to using `"resolve:ssm:*` style syntax. This means that on startup 
+the ASG will read the latest AMI from that location, there's no need to apply TF to update the latest image.
+
+> [!WARNING]
+> User-data may need changing as this uses AL2023 and previous used and AL2.
+>
+> https://docs.aws.amazon.com/AmazonECS/latest/developerguide/al2-to-al2023-ami-transition.html